bug(fix): Implement feedback of product
- Separates components into container and functional
- Solves bug on password numbers regarding user signup

[Fixes #167979793]
