+++ src/pages/Signup/schema.js
@@ -20,7 +20,7 @@ const SignupSchema = Yup.object().shape({
     .trim()
     .min(6, 'minumum of 6 characters')
     .matches(
-      /^(?=.{10,}$)(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*\W).*$/,
+      /^(?=.{8,}$)(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*\W).*$/,
       'Password must be 8 characters long with atleast one special character and uppercase letter'
     )
     .required('Required'),
